pyproject.toml
[tool.poetry]
name = "xoadmin"
version = "1.2.0"
description = "bindings and wrappers to manage an XOA instance"
authors = ["dennis <it.admin@elnissi.co.id>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{ include = "xoadmin", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9.2"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
httpx = "^0.27.0"
colorlog = "^6.8.2"
websockets = "^12.0"
pydantic = "^2.6.4"
click = "^8.1.7"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
xoadmin = "xoadmin.cli.cli:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-httpx = "^0.30.0"
flake8 = "^7.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
